@@ -38,7 +38,7 @@ dependencies = [
 ]
 license = {text = "AGPL"}
 requires-python = ">=3.8"
-version = "1.3.0"
+version = "1.4.0"
 
 [project.optional-dependencies]
 pyarrow = ["pyarrow>=16.0.0,<17.0.0"]
@@ -64,7 +64,8 @@ find = {}
 # preview = true
 
 [tool.ruff]
-target-version = "py38"
+target-version = "py39"
+
 select = [
     "B",   # likely bugs and design problems
     "BLE", # do not catch blind exception
@@ -92,7 +93,6 @@ select = [
 ]
 ignore = [
     "E712",    # comparison to bool: violated by SQLAlchemy filters
-    "PT004",   # fixture does not return anything, add leading underscore: violated by clean_db
     "PLC1901", # simplify comparison to empty string: violated by SQLAlchemy filters
     "TRY003",  # allow providing an error message to exception instance
 ]
@@ -101,6 +101,7 @@ ignore = [
 "ckanext/charts/tests*" = ["S", "PL"]
 "ckanext/charts/logic/schema.py" = ["PLR0913"]
 "ckanext/charts/chart_builders/plotly.py" = ["C901", "PLR0912"]
+"mkdocstrings_handlers/ChartFieldsHandler/__init__.py" = ["N999"] # allow not snake-case module name
 
 [tool.isort]
 known_ckan = "ckan"
@@ -116,7 +117,7 @@ filterwarnings = [
 ]
 
 [tool.pyright]
-pythonVersion = "3.8"
+pythonVersion = "3.9"
 include = ["ckanext"]
 exclude = ["**/test*", "**/migration"]
 strict = []
