@@ -11,4 +11,4 @@ SUPPORTED_CACHE_STRATEGIES = [
 ]
 
 REDIS_PREFIX = "ckanext-charts:*"
-CHART_DEFAULT_ROW_LIMIT = 100
+CHART_DEFAULT_ROW_LIMIT = 1000
